Fixes a start up error with devices attached
If the launchpad and audio interface are both connected at startup,
the jack server failed to start. The solution was to reorder the
refresh sequence at during start up, so that udev goes first and
initialises the jack server before the sequencer enumeration tries to
create the Launchpad controller object.